Failed attempt to fix text alpha
I tried to create an OutlineLabelConvert that automatically changes a
label to a CanvasGroup and sets the correct stuff.
There are several shortcomings:
1.This fucks up the node paths and creates crashes
2.CanvasGroup doesn't work well when parents have modulate, so to fix
  that I had to use top_level and recalculate self_modulate everytime on
  process, which is a bad solution. (possibly very inefficient)
3.Even with the (2) hacked around, if you manually change the CanvasGroup
  you have to be careful to change the self_modulate and not modulate,
  creating many other changes necessary in the code.

Using a OutlineLabelConvert is clearly a bad idea. Just changing the
labels to a instance of OutlineLabel is better, but (3) would still be a
problem and we would need to change everything that changes modulate.